frontend/requirements.txt
google-api-python-client
flask
google-auth
google-auth-oauthlib
google-auth-httplib2
pymongo
bs4
requests
python-dotenv